Fix directory names in blade Fixed Tautology and Contradiction bug. Still need to implement class
WIP Experience

+++ blossom/updates/create_milestones_table.php
@@ -13,14 +13,19 @@ class CreateMilestonesTable extends Migration
             $table->engine = 'InnoDB';
             $table->increments('id');
             $table->string('name');
+            $table->integer('experience_id')->unsigned()->index();
             $table->integer('points');
             $table->timestamps();
+            $table->foreign('experience_id')->references('id')->on('delphinium_blossom_experiences');
         });
     }
 
     public function down()
     {
-        Schema::dropIfExists('delphinium_blossom_milestones');
+        if ( Schema::hasTable('delphinium_blossom_milestones') )
+        {
+            Schema::dropIfExists('delphinium_blossom_milestones');
+        }
     }
 
 }
